refactor(lotto): lottoNumber 분리
