+++ src/main/kotlin/lotto/domain/LottoNumber.kt
@@ -1,10 +1,9 @@
 package lotto.domain
 
 @JvmInline
-value class LottoNumbers(val lottoNumbers: List<Int>) {
+value class LottoNumber(val number: Int) {
     init {
-        require(lottoNumbers.count() == SIZE)
-        if (!lottoNumbers.all { it in MIN..MAX }) {
+        if (number !in MIN..MAX) {
             throw IllegalArgumentException("로또 번호는 1~45까지 입력 가능합니다.")
         }
     }
@@ -12,6 +11,5 @@ value class LottoNumbers(val lottoNumbers: List<Int>) {
     companion object {
         private const val MIN = 1
         private const val MAX = 45
-        private const val SIZE = 6
     }
 }
