+++ src/main/kotlin/lotto/dto/LottoNumbers.kt
@@ -1,0 +1,7 @@
+package lotto.dto
+
+import lotto.domain.LottoNumber
+
+class LottoNumbers(numbers: List<Int>) {
+    val lottoNumbers: List<LottoNumber> = numbers.map { LottoNumber(it) }
+}
