+++ src/main/kotlin/lotto/util/AutoNumbers.kt
@@ -1,7 +1,7 @@
 package lotto.util
 
-import lotto.domain.LottoNumbers
+import lotto.domain.LottoTicket
 
 interface AutoNumbers {
-    fun generateNumbers(): LottoNumbers
+    fun generateNumbers(): LottoTicket
 }
