+++ src/main/kotlin/lotto/util/RandomNumbers.kt
@@ -1,13 +1,14 @@
 package lotto.util
 
-import lotto.domain.LottoNumbers
+import lotto.domain.LottoTicket
+import lotto.dto.LottoNumbers
 
 object RandomNumbers : AutoNumbers {
     private const val BOUND = 45
     private const val SIZE = 6
     private val LOTTO_NUMBERS: List<Int> = List(BOUND) { it + 1 }
 
-    override fun generateNumbers(): LottoNumbers {
-        return LottoNumbers(LOTTO_NUMBERS.shuffled().take(SIZE))
+    override fun generateNumbers(): LottoTicket {
+        return LottoTicket(LottoNumbers(LOTTO_NUMBERS.shuffled().take(SIZE)))
     }
 }
